moved config to package. added test for config
