+++ src/service/main.go
@@ -3,20 +3,20 @@ package main
 import (
 	"fmt"
 	"net/http"
-	"os"
 
 	"github.com/labstack/echo"
 	"github.com/labstack/echo/middleware"
 
 	"github.com/Foundation-13/mwarehouse/src/service/api"
 	"github.com/Foundation-13/mwarehouse/src/service/aws"
+	"github.com/Foundation-13/mwarehouse/src/service/config"
 	"github.com/Foundation-13/mwarehouse/src/service/db"
 	"github.com/Foundation-13/mwarehouse/src/service/storage"
 	"github.com/Foundation-13/mwarehouse/src/service/utils"
 )
 
 func main() {
-	cfg, err := FromEnvironment()
+	cfg, err := config.FromEnvironment()
 	if err != nil {
 		panic(err)
 	}
@@ -45,26 +45,4 @@ func main() {
 	})
 
 	e.Logger.Fatal(e.Start(":8765"))
-}
-
-type Config struct {
-	Region 			string
-	TempBucketName 	string
-}
-
-func FromEnvironment () (Config, error) {
-	region := os.Getenv("REGION")
-	if region == "" {
-		return Config{}, fmt.Errorf("REGION env variable is not defined")
-	}
-
-	tempBucketName := os.Getenv("TEMP_BUCKET_NAME")
-	if tempBucketName == "" {
-		return Config{}, fmt.Errorf("TEMP_BUCKEY_NAME env variable is not defined")
-	}
-
-	return Config{
-		Region: region,
-		TempBucketName: tempBucketName,
-	}, nil
 }
